+++ src/components/common/Header/Header.tsx
@@ -1,31 +1,19 @@
 import React, { useState, useEffect, useRef } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import { useAppDispatch, useAppSelector } from "../../../hooks";
-import { logout, getAuthState } from "../../../features/loginSlice";
-import { openModal, selectOpenState } from "../../../features/modalSlice";
+import { useAppDispatch } from "../../../hooks";
 import { fetchSearchProducts, searchReset } from "../../../features/searchSlice";
-import { NormalBtn } from "../Button/Button";
-import ArrowModal from "../ArrowModal/ArrowModal";
-import Modal from "../Modal/Modal";
+import IconNav from "./IconNav";
 
-import { ReactComponent as CartIcon } from "../../../assets/icons/icon-shopping-cart.svg";
-import { ReactComponent as UserIcon } from "../../../assets/icons/icon-user.svg";
 import logo from "../../../assets/icons/Logo-hodu.svg";
 import searchIcon from "../../../assets/icons/search.svg";
-import shoppingBag from "../../../assets/icons/icon-shopping-bag.svg";
 import * as S from "./headerStyle";
 
 export function Header() {
   const navigate = useNavigate();
   const { pathname } = useLocation();
-
   const dispatch = useAppDispatch();
-  const { userName, token, userType } = useAppSelector(getAuthState);
 
   const inputRef = useRef() as React.MutableRefObject<HTMLInputElement>;
-  const modal = useAppSelector(selectOpenState);
-
-  const [onArrowModal, setArrowModal] = useState(false);
   const [searchContent, setSearchContent] = useState("");
 
   useEffect(() => {
@@ -46,31 +34,8 @@ export function Header() {
     navigate(`/search/${searchContent}`);
   };
 
-  //로그아웃
-  const onLogout = () => {
-    setArrowModal(false);
-    dispatch(logout());
-    navigate("/");
-  };
-
-  const needLoginModal = (
-    <Modal onClickYes={() => navigate("/login")}>
-      로그인이 필요한 서비스 입니다. <br /> 로그인 하시겠습니까?
-    </Modal>
-  );
-
-  //모달 정보
-  const arrowList = [
-    { label: "마이페이지", onClick: () => navigate("/mypage") },
-    { label: "로그아웃", onClick: () => onLogout() },
-  ];
-
-  //헤더 아이콘 색상 변경
-  const onArrowIcon = !token ? undefined : onArrowModal ? "open" : "close";
-
   return (
     <>
-      {!token && modal ? needLoginModal : null}
       <S.HeaderContainer>
         <S.HeaderContents>
           <S.LeftWrap>
@@ -86,43 +51,7 @@ export function Header() {
             </S.SearchForm>
           </S.LeftWrap>
           <S.RightWrap>
-            {userType === "SELLER" ? (
-              <>
-                <NormalBtn
-                  onClick={() => navigate("/center")}
-                  icon={shoppingBag}
-                  width="168px"
-                  padding="10px 20px"
-                >
-                  판매자 센터
-                </NormalBtn>
-                <S.UerModalWrap>
-                  <S.UserBtn onClick={() => setArrowModal(!onArrowModal)} arrow={onArrowIcon}>
-                    <UserIcon stroke={"black"} />
-                    <small>{token ? userName : "로그인"}</small>
-                  </S.UserBtn>
-                  <ArrowModal on={onArrowModal} list={arrowList} />
-                </S.UerModalWrap>
-              </>
-            ) : (
-              <>
-                <S.NavBtn
-                  onClick={token ? () => navigate("/cart") : () => dispatch(openModal("예"))}
-                >
-                  <CartIcon stroke={"black"} />
-                </S.NavBtn>
-                <S.UerModalWrap>
-                  <S.UserBtn
-                    onClick={token ? () => setArrowModal(!onArrowModal) : () => navigate("/login")}
-                    arrow={onArrowIcon}
-                  >
-                    <UserIcon stroke={"black"} />
-                    <small className="txt-ellipsis">{token ? userName : "로그인"}</small>
-                  </S.UserBtn>
-                  <ArrowModal on={onArrowModal} list={arrowList} />
-                </S.UerModalWrap>
-              </>
-            )}
+            <IconNav />
           </S.RightWrap>
         </S.HeaderContents>
       </S.HeaderContainer>
@@ -130,6 +59,7 @@ export function Header() {
   );
 }
 
+// 판매자 센터 헤더
 export function CenterHeader() {
   const navigate = useNavigate();
